Fix tp command for invalid input

+++ core/src/no/erlyberly/bootlegterraria/world/TiledGameMap.java
@@ -35,7 +35,7 @@ public class TiledGameMap extends GameMap {
         try {
             this.tiledMap = new TmxMapLoader().load(map);
         } catch (final Exception e) {
-            GameMain.getConsoleHandler().log("Failed to load map '%s'", LogLevel.ERROR, map);
+            GameMain.getConsoleHandler().logf("Failed to load map '%s'", LogLevel.ERROR, map);
         }
 
         MapProperties mapProperties = this.tiledMap.getProperties();
